Clarify that bounding boxes are insets
The nomenclature was confusing. The so-called "bounding box" was
actually a set of insets in LTRB order, which could be applied to any
rectangular space to create corresponding bounds within that space.

The methods were confusingly named, too: asking a box for its box is
not something we want to be encouraging. ;-)